[OPP-1379] ryddet i prosjektstrukturen
