+++ .nais/preprod.yaml
@@ -58,6 +58,8 @@ spec:
       value: "https://modapp-q1.adeo.no/henvendelse/services/domene.Virksomhet/Dialog_v1"
     - name: SF_HENVENDELSE_URL
       value: ""
+    - name: KODEVERK_URL
+      value: ""
     - name: PDL_URL
       value: "https://pdl-api.nais.preprod.local/graphql"
     - name: UNLEASH_API_URL
