+++ app/src/main/kotlin/no/nav/henvendelse/config/CxfConfig.kt
@@ -1,6 +1,6 @@
 package no.nav.henvendelse.config
 
-import no.nav.henvendelse.consumer.unleash.UnleashService
+import no.nav.henvendelse.service.unleash.UnleashService
 import no.nav.henvendelse.service.dialog.HenvendelseDialogService
 import no.nav.henvendelse.service.dialog.SfDialogService
 import no.nav.henvendelse.soap.DialogWs
