+++ app/src/main/kotlin/no/nav/henvendelse/config/PdlConfig.kt
@@ -1,8 +1,9 @@
-package no.nav.henvendelse.consumer.pdl
+package no.nav.henvendelse.config
 
 import io.ktor.util.*
 import no.nav.common.rest.client.RestClient
 import no.nav.common.sts.SystemUserTokenProvider
+import no.nav.henvendelse.service.pdl.PdlService
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
