+++ app/src/main/kotlin/no/nav/henvendelse/config/SfDialogConfig.kt
@@ -1,15 +1,15 @@
-package no.nav.henvendelse.consumer.sfhenvendelse
+package no.nav.henvendelse.config
 
 import no.nav.common.health.selftest.SelfTestCheck
 import no.nav.common.rest.client.RestClient
 import no.nav.common.sts.SystemUserTokenProvider
 import no.nav.common.utils.EnvironmentUtils
-import no.nav.henvendelse.consumer.kodeverk.KodeverkService
-import no.nav.henvendelse.consumer.pdl.PdlService
 import no.nav.henvendelse.consumer.sfhenvendelse.generated.apis.HenvendelseInfoApi
 import no.nav.henvendelse.consumer.sfhenvendelse.generated.apis.KodeverkApi
 import no.nav.henvendelse.log
 import no.nav.henvendelse.service.dialog.SfDialogService
+import no.nav.henvendelse.service.kodeverk.KodeverkService
+import no.nav.henvendelse.service.pdl.PdlService
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import java.lang.StringBuilder
@@ -45,8 +45,8 @@ class SfDialogConfig {
 
         return SfDialogService(
             HenvendelseInfoApi(url, httpClient),
-            pdlService,
             KodeverkApi(url, httpClient),
+            pdlService,
             kodeverkService
         )
     }
