+++ app/src/main/kotlin/no/nav/henvendelse/soap/DialogWs.kt
@@ -1,6 +1,6 @@
 package no.nav.henvendelse.soap
 
-import no.nav.henvendelse.consumer.unleash.UnleashService
+import no.nav.henvendelse.service.unleash.UnleashService
 import no.nav.henvendelse.service.dialog.DialogV1Service
 import no.nav.henvendelse.service.dialog.HenvendelseDialogService
 import no.nav.henvendelse.service.dialog.SfDialogService
